@@ -3,22 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.3.0",
+    "@craco/craco": "^6.4.3",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.60",
-    "axios": "^0.21.1",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^12.1.4",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.26.1",
     "jwt-decode": "^3.1.2",
     "moment": "^2.29.1",
-    "prop-types": "^15.8.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "prop-types": "^15.8.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "sanitize-html": "^2.5.1"
+    "react-scripts": "^4.0.3",
+    "sanitize-html": "^2.7.0",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "concurrently": "^5.3.0",
+    "concurrently": "^7.0.0",
     "prettier": "^2.6.1"
   },
   "scripts": {
@@ -29,7 +33,8 @@
     "server": "npm run start:dev --prefix backend",
     "dev": "concurrently \"npm run server\" \"npm start\"",
     "postinstall": "cd backend && npm install",
-    "lint": "npx prettier --write ."
+    "lint": "npx prettier --check .",
+    "lint:format": "npx prettier --write ."
   },
   "eslintConfig": {
     "extends": "react-app"
