@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
@@ -17,15 +17,15 @@ const theme = createTheme({
   },
 });
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root"));
+root.render(
   <ThemeProvider theme={theme}>
     <AuthContextProvider>
       <NowPlayingContextProvider>
         <App />
       </NowPlayingContextProvider>
     </AuthContextProvider>
-  </ThemeProvider>,
-  document.getElementById("root")
+  </ThemeProvider>
 );
 
 // If you want your app to work offline and load faster, you can change
